move TestHUPGood2Alt to a flake suite
